+++ CHANGELOG.md
@@ -20,6 +20,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix 1600: Replicas fragment not set 
 * Fix 1579: Added documentation for clearing doubts related to sidecars
 * Fix 1587: HandlerHub refactored
+* Fix 1579: Add ImageChange triggers to application container when resource fragments are used
 
 ### 4.0.0 (14-03-2019)
 * Refactor 678: Enricher Workflow to allow direct generation of OpenShift objects.